Weaponised Wheelchair Bells (#90176)
This PR does two things.

Thing one:
- Ringing the bell on your motorized wheelchair while it is emagged has
a 20% chance to detonate the bomb.

Thing two:
- You can attach TTVs to wheelchairs, which are detonated either by
ringing the bell repeatedly until it explodes or (if emagged) running
people over.

![image](https://github.com/user-attachments/assets/bb52d077-17f3-49cf-a267-5e6a5d61f9ea)

You can remove the TTV by deconstructing the wheelchair with a wrench.
Also I did some generalised code cleanup in the file, while I was there,
and fixed a bug with emagging actually only working if the panel was
closed (and complaining it was closed if it was open).

It provides a novel and amusing means of payload delivery which it is
almost impossible for the perpetrator to survive.
Wheelchair owners may use this as an effective deterrent to ensure that
nobody is pushing their chair unasked.

Also the tg host asked me to do it because he saw it on TV.

:cl:
add: You can now add bell-detonated TTVs to wheelchairs.
add: Ringing the bell on an emagged wheelchair may cause it to explode.
/:cl: